Overhaul of how command line tokenization works
There were a lot of issues with how the command line was parsed
and tokenized, specifically variable expansion would happen first,
the the command line would be parsed, but this lead to some strangeness
that doesn't happen in bash (like redirection happening in a variable).

This moves all variable expansion logic and backtick processing
logic into the tokenizer, so it can make the decision, token by token,
on whether or not variables get expanded and how that expansion takes
place. I won't go into the details of what this addressed, but
you can look at the newly added `TokenizerTest` (yay, tests!) to get
an idea of the behavior.